+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Add custom envelopes for benjee10's SOCK
+* Add `transition_offset` to the planet pack configs
 
 ## 1.0.0.0 - 2025-02-05